rag_queue/requirements.txt
rq
redis
fastapi
uvicorn
requests
python-dotenv
langchain-huggingface
langchain-qdrant
qdrant-client
google-generativeai
sentence-transformers
